(refactor) auth files to fit into new file storage structure
